fix regression with boost fusion 1.58
see commit 07a7e21bfb76dfe2cb2472c13fcf6fb1c7367748.

+++ dune/dpg/radiative_transfer/scattering.hh
@@ -160,7 +160,7 @@ assembleScattering(BlockVector<FieldVector<double,1> >& scattering,
 
   globalTotalTestSize =
     fold(zip(globalTestSpaceOffsets, testBasisIndexSet),
-         0, globalOffsetHelper());
+         (size_t)0, globalOffsetHelper());
 
   if(!isSaddlepoint)
   {
